SpecPlay2022: Gave ability to swap convos simply by clicking on a character
After playtesting, I realized that having to click "x" between each convo to then switch to another npc was too long a process. My inclination was to be able to switch convos quickly by clicking on another npc (without having to go through the intermediary step of clicking "x" on the keyboard), so I coded that option!