SpecPlay2022: fixed keyboard position; added hover styles to left-right-buttons and exit-button
My playtesters mentioned that sometimes the keyboard does not fit on the screen and they need to reduce their window size to access it.

My solution is that I ensured the keyboard is always visible by locking its position to the bottom of the screen so that it does not overflow off-screen.